% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cars-data.R
\docType{data}
\name{cars}
\alias{cars}
\title{Stated car choice observations by US and Chinese car buyers}
\format{
\tabular{ll}{
   Variable \tab Description \cr
}
}
\source{
Raw data downloaded from \href{https://github.com/jhelvy/tra2015}{this repo}
}
\usage{
data(cars)
}
\description{
Data from Helveston et al. (2015) containing 448 stated choice observations
from Chinese car buyers and 384 stated choice observations from US car
buyers. Conjoint surveys were fielded in 2012 in four major Chinese cities
(Beijing, Shanghai, Shenzhen, and Chengdu), online in the US on Amazon
Mechanical Turk, and in person at the Pittsburgh Auto show. Participants
were asked to select a vehicle from a set of three alternatives. Each
participant answered 15 choice questions.
}
\examples{
data(cars)

head(cars)
}
\references{
Helveston, J. P., Liu, Y., Feit, E. M., Fuchs, E. R. H.,
Klampfl, E., & Michalek, J. J. (2015). "Will Subsidies Drive Electric
Vehicle Adoption? Measuring Consumer Preferences in the U.S. and China."
Transportation Research Part A: Policy and Practice, 73, 96–112.
https://doi.org/10.1016/j.tra.2015.01.002
}
\keyword{datasets}
