% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{coef.logitr}
\alias{coef.logitr}
\title{Returns the coefficients of an estimated model of the 'logitr' class.}
\usage{
\method{coef}{logitr}(model)
}
\description{
Returns the coefficients of an estimated model of the 'logitr' class.
}
\examples{
# Get the coefficients of an estimated model:
coef(model)
}
\keyword{coef}
\keyword{logitr,}
