% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelInputs.R
\name{dummyCode}
\alias{dummyCode}
\title{Creates dummy-coded variables.}
\usage{
dummyCode(df, vars)
}
\arguments{
\item{df}{A data frame.}

\item{vars}{The variables in the data frame for which you want to
create new dummy coded variables.}
}
\description{
Use this function to create dummy-coded variables in a data frame.
}
\examples{
# Create an example data frame:
df <- data.frame(
    animal   = c("dog", "goldfish", "bird", "dog", "goldfish"),
    numLegs  = c(4, 0, 2, 4, 0),
    lifeSpan = c(10, 10, 5, 10, 10))

# Create dummy coded variables for the variables "animal" and "numLegs":
df_dummy <- dummyCode(df, vars = c("animal", "numLegs"))
df_dummy
}
