% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.logitr}
\alias{fitted.logitr}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{logitr}(object)
}
\arguments{
\item{object}{is an object of class \code{logitr}.}
}
\value{
A data frame of the \code{obsID} and the fitted values extracted from
\code{object}.
}
\description{
Returns fitted values from an object of class \code{logitr}.
}
\examples{
library(logitr)

# Estimate a preference space model
mnl_pref <- logitr(
  data   = yogurt,
  choice = "choice",
  obsID  = "obsID",
  pars   = c("price", "feat", "brand")
)

# Extract the fitted values from the model
fitted(mnl_pref)
}
\keyword{fitted}
\keyword{fitted.values}
\keyword{logitr}
