% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getCoefTable}
\alias{getCoefTable}
\title{Get the coefficient summary table as a data frame}
\usage{
getCoefTable(object)
}
\arguments{
\item{x}{is an object of class \code{logitr}.}
}
\value{
Returns a data frame of the coefficient summary table of a model
estimated using the \code{logitr()} function.
}
\description{
Returns a data frame of the coefficient summary table for a \code{logitr} class
object.
}
\examples{
library(logitr)

# Run a MNL model in the preference space
mnl_pref <- logitr(
  data = yogurt,
  choiceName = "choice",
  obsIDName = "obsID",
  parNames = c("price", "feat", "dannon", "hiland", "yoplait")
)

# Get the coefficient summary table as a data frame
getCoefTable(mnl_pref)
}
