% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRLS.R
\name{gradient_IRLS_logistic}
\alias{gradient_IRLS_logistic}
\title{Gradient of IRLS Logistic Regression}
\usage{
gradient_IRLS_logistic(betas, X, y, lam = 0, weights = NULL,
  penalty = "L2", vec = NULL)
}
\arguments{
\item{betas}{beta estimates (includes intercept)}

\item{X}{matrix or data frame}

\item{y}{response vector of 0,1}

\item{lam}{tuning parameter for regularization term}

\item{weights}{optional vector of weights for IRLS}

\item{penalty}{choose from BLAH. Defaults to 'L2'}

\item{vec}{optional vector to specify which coefficients will be penalized}
}
\value{
returns the gradient
}
\description{
Computes the gradient of logistic regression. We use this to determine if the KKT conditions are satisfied. This function is to be used with the 'IRLS' function.
}
\examples{
gradient_IRLS_logistic(betas, X, y, lam = 0.1, penalty = 'L2')
}

