% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{gradient_linear}
\alias{gradient_linear}
\title{Gradient of Linear Regression}
\usage{
gradient_linear(betas, X, y, lam = 0, weights = NULL, penalty = "None",
  vec)
}
\arguments{
\item{betas}{beta estimates (includes intercept)}

\item{X}{matrix or data frame}

\item{y}{response vector of 0,1}

\item{lam}{tuning parameter for regularization term}

\item{weights}{option vector of weights for weighted least squares}

\item{penalty}{choose from c('None', 'Ridge'). Defaults to 'None'}

\item{vec}{vector to specify which coefficients will be penalized}
}
\value{
returns the gradient
}
\description{
Computes the gradient of linear regression (optional Ridge regularization term). This function is to be used with the 'Linearr' function.
}
\examples{
gradient_linear(betas, X, y, lam = 0.1, penalty = 'Ridge')
}

