% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRLS.R
\name{IRLS}
\alias{IRLS}
\title{Iterative Re-Weighted Least Squares}
\usage{
IRLS(X, y, lam = 0, weights = NULL, penalty = "L2", tol = 10^(-5),
  maxit = 1e+05, vec)
}
\arguments{
\item{X}{matrix or data frame}

\item{y}{matrix or vector of response 0,1}

\item{lam}{tuning parameter for regularization term}

\item{weights}{optional vector of weights for IRLS}

\item{penalty}{choose from BLAH. Defaults to 'L2'}

\item{tol}{tolerance - used to determine algorithm convergence}

\item{maxit}{maximum iterations}

\item{vec}{optional vector to specify which coefficients will be penalized}
}
\value{
returns beta estimates (includes intercept), total iterations, and gradients.
}
\description{
Computes the logistic regression coefficient estimates using the iterative re-weighted least squares (IRLS) algorithm. This function is to be used with the 'logistic' function.
}
\examples{
IRLS(X, y, n.list = c(rep(1, n)), lam = 0.1, alpha = 1.5)
}

