% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{linear}
\alias{linear}
\title{Linear}
\usage{
linear(X, y, lam = 0, weights = NULL, intercept = TRUE, vec = NULL)
}
\arguments{
\item{X}{matrix or data frame}

\item{y}{matrix or data frame of response values}

\item{lam}{optional tuning parameter for L2 regularization term. Defaults to `lam = 0`}

\item{weights}{optional vector of weights for weighted least squares}

\item{intercept}{add column of ones if not already present. Defaults to TRUE}

\item{vec}{optional vector to specify which coefficients will be penalized}
}
\value{
returns the coefficient estimates
}
\description{
Computes the linear regression coefficient estimates (L2-penalization and weights, optional)
}
\examples{
library(dplyr)
X = dplyr::select(iris, -c(Species, Sepal.Length))
y = dplyr::select(iris, Sepal.Length)
linear(X, y, lam = 0.1, weights = rep(1, 150))
}

