% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic}
\alias{logistic}
\title{Logistic Regression}
\usage{
logistic(X, y, lam = 0, penalty = "L2", intercept = TRUE,
  method = "IRLS", tol = 10^(-5), maxit = 10^(5), vec = NULL)
}
\arguments{
\item{X}{matrix or data frame}

\item{y}{matrix or vector of response values 0,1}

\item{lam}{optional tuning parameter for L2 regularization term. Defaults to `lam = 0`}

\item{penalty}{choose from BLAH. Defaults to 'L2'}

\item{intercept}{Defaults to TRUE}

\item{method}{optimization algorithm. Choose from 'IRLS' or 'MM'. Defaults to 'IRLS' if BLAH}

\item{tol}{tolerance - used to determine algorithm convergence. Defaults to 10^-5}

\item{maxit}{maximum iterations. Defaults to 10^5}

\item{vec}{optional vector to specify which coefficients will be penalized}
}
\value{
returns the coefficient estimates
}
\description{
Computes the coefficient estimates for logistic regression. Regularization penalty optional.
}
\examples{
library(dplyr)
X = dplyr::select(iris, -c(Species, Sepal.Length))
y = dplyr::select(iris, Sepal.Length)
logistic(X, y, lam = 0.1)
}

