% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitr-package.R
\docType{package}
\name{logitr-package}
\alias{logitr-package}
\title{logitr: Logit Models w/Preference & WTP Space Utility Parameterizations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Fast estimation of multinomial logit (MNL) and mixed logit (MXL) models. Models can be estimated using preference space or willingness-to-pay (WTP) space utility parameterizations. Weighted models can also be estimated. An option is available to run a parallelized multistart optimization loop with random starting points in each iteration, which is useful for non-convex problems like MXL models or models with WTP space utility parameterizations. The main optimization loop uses the 'nloptr' package to minimize the negative log-likelihood function. Additional functions are available for computing and comparing WTP from both preference space and WTP space models and for predicting expected choices and choice probabilities for sets of alternatives based on an estimated model. Mixed logit models can include uncorrelated or correlated heterogeneity covariances and are estimated using maximum simulated likelihood based on the algorithms in Train (2009) \doi{10.1017/CBO9780511805271}. More details can be found in Helveston (2023) \doi{10.18637/jss.v105.i10}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jhelvy/logitr}
  \item \url{https://jhelvy.github.io/logitr/}
  \item Report bugs at \url{https://github.com/jhelvy/logitr/issues}
}

}
\author{
\strong{Maintainer}: John Helveston \email{john.helveston@gmail.com} (\href{https://orcid.org/0000-0002-2657-9191}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Connor Forsythe \email{cforsyth@andrew.cmu.edu} [contributor]
}

}
\keyword{internal}
