% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{logitr.coefTable}
\alias{logitr.coefTable}
\title{Returns the summary coefficient table of an estimated model of the
'logitr' class.}
\usage{
logitr.coefTable(model)
}
\description{
Returns the summary coefficient table of an estimated model of the
'logitr' class.
}
\examples{
# Get the coefficients table of an estimated model:
logitr.coefTable(model)
}
\keyword{coefTable}
\keyword{logitr,}
