% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{logitr.statTable}
\alias{logitr.statTable}
\title{Returns the summary statistics table of an estimated model of the
'logitr' class.}
\usage{
logitr.statTable(model)
}
\description{
Returns the summary statistics table of an estimated model of the
'logitr' class.
}
\examples{
# Get the statistics table of an estimated model:
logitr.statTable(model)
}
\keyword{logitr,}
\keyword{statTable}
