% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{logitr.statusCodes}
\alias{logitr.statusCodes}
\title{Prints the status codes from the nloptr optimization routine.}
\usage{
logitr.statusCodes()
}
\description{
Prints the status codes from the nloptr optimization routine.
}
\examples{
# View the status codes:
logitr.statusCodes()
}
\keyword{codes}
\keyword{logitr,}
\keyword{nloptr,}
\keyword{status}
