% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{logitr.summary}
\alias{logitr.summary}
\title{Prints a summary of an estimated model using the logitr package.}
\usage{
logitr.summary(model)
}
\description{
Prints a summary of an estimated model using the logitr package.
}
\examples{
# View a summary of an estimate model:
logitr.summary(model)
}
\keyword{codes}
\keyword{logitr,}
\keyword{nloptr,}
\keyword{status}
