% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRLS.R
\name{loss_logistic}
\alias{loss_logistic}
\title{Loss of Logistic Regression}
\usage{
loss_logistic(betas, X, y, lam = 0, vec)
}
\arguments{
\item{betas}{beta estimates (includes intercept)}

\item{X}{matrix or data frame}

\item{y}{response vector of 0,1}

\item{lam}{tuning parameter for ridge regularization term}

\item{vec}{vector to specify which coefficients will be penalized}
}
\value{
returns the gradient
}
\description{
Computes the loss of logistic regression (optional ridge regularization term). This is used for 'safe' Newton's method. This function is to be used with the 'IRLS' function.
}

