% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.logisticr}
\alias{predict.logisticr}
\title{Predict Logistic Regression}
\usage{
\method{predict}{logisticr}(object, X, y = NULL)
}
\arguments{
\item{object}{'logisticr' object or matrix of betas}

\item{X}{matrix or data frame of (new) observations}

\item{y}{optional, matrix or vector of response values 0,1}
}
\description{
Generates prediction for logistic regression. Note that one can either input a 'logisticr' object or a matrix of beta coefficients.
}
\examples{

fitted = logisticr(X, y, lam = 0.1, penalty = 'ridge', method = 'MM')
predict.logisticr(fitted, X)
}

