% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{print.linearr}
\alias{print.linearr}
\title{Print linearr object}
\usage{
\method{print}{linearr}(x, ...)
}
\arguments{
\item{x}{linearr class object}

\item{...}{additional arguments}
}
\description{
Print linearr object.
}
\keyword{internal}
