% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{print.logisticr}
\alias{print.logisticr}
\title{Print logisticr object}
\usage{
\method{print}{logisticr}(x, ...)
}
\arguments{
\item{x}{logisticr class object}

\item{...}{additional arguments}
}
\description{
Print logisticr object.
}
\keyword{internal}
