% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.logitr_validation}
\alias{print.logitr_validation}
\title{Print method for logitr_validation objects}
\usage{
\method{print}{logitr_validation}(x, ...)
}
\arguments{
\item{x}{An object of class 'logitr_validation'}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for logitr_validation objects
}
