% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRLS.R
\name{softmax}
\alias{softmax}
\title{Softmax}
\usage{
softmax(X, b)
}
\arguments{
\item{X}{matrix (includes intercept)}

\item{b}{beta estimates (includes intercept)}
}
\value{
returns the estimated probability
}
\description{
Computes the probability for each observation in X being in class '1'.
}
\examples{
softmax(X, b)
}

