% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{statusCodes}
\alias{statusCodes}
\title{View a description the nloptr status codes}
\usage{
statusCodes()
}
\value{
No return value; prints a summary of the \code{nloptr} status codes to
the console.
}
\description{
Prints a description of the status codes from the nloptr optimization routine.
}
\examples{
statusCodes()
}
\keyword{codes}
\keyword{logitr}
\keyword{nloptr}
\keyword{status}
