% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{tidy.logitr}
\alias{tidy.logitr}
\title{\code{tidy()} broom method for \code{logitr} class objects.}
\usage{
\method{tidy}{logitr}(x, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{is an object of class \code{logitr}.}

\item{conf.int}{Include a confidence interval?}

\item{...}{further arguments.}

\item{level}{The confidence level required.}
}
\value{
A tibble of the model coefficients and (optional) confidence
intervals.
}
\description{
\code{tidy()} broom method for \code{logitr} class objects.
}
\examples{
library(logitr)

# Estimate a preference space model
mnl_pref <- logitr(
  data    = yogurt,
  outcome = "choice",
  obsID   = "obsID",
  pars    = c("price", "feat", "brand")
)

# Extract a tibble of the model coefficients
tidy(mnl_pref)

# Extract a tibble of the model coefficients with confidence intervals
tidy(mnl_pref, conf.int = TRUE)

}
