% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{wtp}
\alias{wtp}
\title{Returns the computed WTP from a preference space model.}
\usage{
wtp(model, priceName)
}
\description{
Returns the computed WTP from a preference space model.
}
\examples{
# Run a MNL model in the Preference Space:
data(yogurt)

mnl.pref = logitr(
  data       = yogurt,
  choiceName = 'choice',
  obsIDName  = 'obsID',
  parNames   = c('price', 'feat', 'dannon', 'hiland', 'yoplait'))

# Get the WTP implied from the preference space model
wtp(mnl.pref, priceName='price')
}
\keyword{logitr,}
\keyword{wtp}
