% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.logitr}
\alias{coef.logitr}
\title{Returns the coefficients of an estimated model of the 'logitr' class.}
\usage{
\method{coef}{logitr}(model)
}
\description{
Returns the coefficients of an estimated model of the 'logitr' class.
}
\examples{
# Run a MNL model in the Preference Space:
data(yogurt)

mnl.pref = logitr(
  data       = yogurt,
  choiceName = 'choice',
  obsIDName  = 'obsID',
  parNames   = c('price', 'feat', 'dannon', 'hiland', 'yoplait'))

# Get the model coefficients:
coef(model)
}
\keyword{coef}
\keyword{logitr,}
