% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{conditional_predict}
\alias{conditional_predict}
\title{Predict conditional probabilities for known panel individuals}
\usage{
conditional_predict(
  object,
  conditioning_data,
  newdata,
  returnData = FALSE,
  ...
)
}
\arguments{
\item{object}{is an object of class \code{logitr} estimated using \code{logitr()}.}

\item{conditioning_data}{a \code{data.frame} in long format containing the prior
observed choices used to condition on each individual's history. It must
include the model's outcome column, observation ID column, panel ID column,
and all predictor columns used by the fitted model.}

\item{newdata}{a \code{data.frame} in long format containing the new choice tasks
for which to compute conditional probabilities.}

\item{returnData}{If \code{TRUE}, also return the provided \code{newdata}.}

\item{...}{further arguments.}
}
\value{
A data frame of conditional predicted probabilities.
}
\description{
Returns conditional predicted probabilities for new choice tasks using
explicit conditioning histories for known panel individuals in a mixed logit
model.
}
\examples{
library(logitr)

mxl_pref <- readRDS(system.file("extdata", "mxl_pref.Rds", package = "logitr"))

conditioning_data <- subset(yogurt, id \%in\% c(1, 2))
task_template <- subset(
  yogurt,
  obsID == 1,
  select = c("obsID", "id", "alt", "price", "feat", "brand")
)
newdata <- rbind(
  transform(task_template, obsID = 9001, id = 1),
  transform(task_template, obsID = 9002, id = 2)
)

conditional_predict(
  mxl_pref,
  conditioning_data = conditioning_data,
  newdata = newdata
)
}
