% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{dummyCode}
\alias{dummyCode}
\title{Add dummy-coded variables to data frame.}
\usage{
dummyCode(df, vars)
}
\arguments{
\item{df}{A data frame.}

\item{vars}{The variables in the data frame for which you want to
create new dummy coded variables.}
}
\value{
A a dataframe with new dummy-coded variables added.
}
\description{
This function is depreciated. Use fastDummies::dummy_cols() instead.
}
\keyword{internal}
