% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{dummyCode}
\alias{dummyCode}
\title{Adds dummy-coded variables to data frame.}
\usage{
dummyCode(df, vars)
}
\arguments{
\item{df}{A data frame.}

\item{vars}{The variables in the data frame for which you want to
create new dummy coded variables.}
}
\value{
A a dataframe with new dummy-coded variables added.
}
\description{
This function adds dummy-coded variables to a data frame based on a
vector of column names.
}
\examples{
# Create an example data frame:
df <- data.frame(
    animal   = c("dog", "goldfish", "bird", "dog", "goldfish"),
    numLegs  = c(4, 0, 2, 4, 0),
    lifeSpan = c(10, 10, 5, 10, 10))

# Create dummy coded variables for the variables "animal" and "numLegs":
df_dummy <- dummyCode(df, vars = c("animal", "numLegs"))
df_dummy
}
