% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{logitr}
\alias{logitr}
\title{The main function for running the logitr program}
\usage{
logitr(data, choiceName, obsIDName, parNames, priceName = NULL,
  randPars = NULL, randPrice = NULL, modelSpace = "pref",
  options = list())
}
\description{
logitr estimates multinomial (MNL) and mixed logit (MXL) models in R.
Models can be estimated in the "Preference Space" or
"Willingness-to-pay (WTP)" space. The program includes many options,
including the option to run a multistart optimization loop with random
starting points in each iteration to search for a global solution, which is
useful for non-convex problems like WTP space models or MXL models.
The main optimization loop uses the nloptr function to minimize the
negative log-likelihood function.
}
\examples{
# For an example go to: https://github.com/jhelvy/logitr/tree/master/example
}
\keyword{logitr,}
\keyword{mnl,}
\keyword{mxl}
