% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitr.R
\name{logitr}
\alias{logitr}
\title{The main function for estimating logit models}
\usage{
logitr(
  data,
  choice,
  obsID,
  pars,
  price = NULL,
  randPars = NULL,
  randPrice = NULL,
  modelSpace = "pref",
  weights = NULL,
  cluster = NULL,
  robust = FALSE,
  numMultiStarts = 1,
  useAnalyticGrad = TRUE,
  scaleInputs = TRUE,
  startParBounds = c(-1, 1),
  standardDraws = NULL,
  numDraws = 50,
  startVals = NULL,
  options = list(print_level = 0, xtol_rel = 1e-06, xtol_abs = 1e-06, ftol_rel = 1e-06,
    ftol_abs = 1e-06, maxeval = 2000, algorithm = "NLOPT_LD_LBFGS")
)
}
\arguments{
\item{data}{The choice data, formatted as a \code{data.frame} object.}

\item{choice}{The name of the column that identifies the choice variable.}

\item{obsID}{The name of the column that identifies each choice
observation.}

\item{pars}{The names of the parameters to be estimated in the model.
Must be the same as the column names in the \code{data} argument. For WTP space
models, do not include price in \code{pars}.}

\item{price}{The name of the column that identifies the price variable.
Required for WTP space models. Defaults to \code{NULL}.}

\item{randPars}{A named vector whose names are the random parameters and
values the distribution: \code{'n'} for normal or \code{'ln'} for log-normal.
Defaults to \code{NULL}.}

\item{randPrice}{The random distribution for the price parameter: \code{'n'} for
normal or \code{'ln'} for log-normal. Only used for WTP space MXL models.
Defaults to \code{NULL}.}

\item{modelSpace}{Set to \code{'wtp'} for WTP space models. Defaults to \code{"pref"}.}

\item{weights}{The name of the column that identifies the weights to be
used in model estimation. Defaults to \code{NULL}.}

\item{cluster}{The name of the column that identifies the cluster
groups to be used in model estimation. Optional. Defaults to \code{NULL}.}

\item{robust}{Determines whether or not a robust covariance matrix is
estimated. Defaults to \code{FALSE}. Specification of a \code{cluster} or
\code{weights} will override the user setting and set this to `TRUE' (a
warning will be displayed in this case). Replicates the functionality of
Stata's cmcmmixlogit.}

\item{numMultiStarts}{is the number of times to run the optimization loop,
each time starting from a different random starting point for each parameter
between \code{startParBounds}. Recommended for non-convex models, such as WTP
space models and mixed logit models. Defaults to \code{1}.}

\item{useAnalyticGrad}{Set to \code{FALSE} to use numerically approximated
gradients instead of analytic gradients during estimation. For now, using
the analytic gradient is faster for MNL models but slower for MXL models.
Defaults to \code{TRUE}.}

\item{scaleInputs}{By default each variable in \code{data} is scaled to be
between 0 and 1 before running the optimization routine because it usually
helps with stability, especially if some of the variables have very large or
very small values (e.g. \verb{> 10^3} or \verb{< 10^-3}). Set to \code{FALSE} to turn this
feature off. Defaults to \code{TRUE}.}

\item{startParBounds}{sets the \code{lower} and \code{upper} bounds for the starting
parameters for each optimization run, which are generated by
\code{runif(n, lower, upper)}. Defaults to \code{c(-1, 1)}.}

\item{standardDraws}{By default, a new set of standard normal draws are
generated during each call to \code{logitr} (the same draws are used during each
multistart iteration). The user can override those draws by providing a
matrix of standard normal draws if desired. Defaults to \code{NULL}.}

\item{numDraws}{The number of Halton draws to use for MXL models for the
maximum simulated likelihood. Defaults to \code{50}.}

\item{startVals}{is vector of values to be used as starting values for the
optimization. Only used for the first run if \code{numMultiStarts > 1}. Defaults
to \code{NULL}.}

\item{options}{A list of options for controlling the \code{nloptr()} optimization.
Run \code{nloptr::nloptr.print.options()} for details.}
}
\value{
The function returns a list object containing the following objects.\tabular{ll}{
   Value \tab Description \cr
   \code{coef} \tab The model coefficients at convergence. \cr
   \code{standErrs} \tab The standard errors of the model coefficients at convergence. \cr
   \code{logLik} \tab The log-likelihood value at convergence. \cr
   \code{nullLogLik} \tab The null log-likelihood value (if all coefficients are 0). \cr
   \code{gradient} \tab The gradient of the log-likelihood at convergence. \cr
   \code{hessian} \tab The hessian of the log-likelihood at convergence. \cr
   \code{covariance} \tab The covariance matrix at convergence. \cr
   \code{numObs} \tab The number of observations. \cr
   \code{numParams} \tab The number of model parameters. \cr
   \code{call} \tab The matched call to \code{logitr()}. \cr
   \code{inputs} \tab A list of the original inputs to \code{logitr()}. \cr
   \code{freq} \tab The frequency counts of each choice alternative. \cr
   \code{startPars} \tab The starting values used. \cr
   \code{multistartNumber} \tab The multistart run number for this model. \cr
   \code{multistartSummary} \tab A summary of the log-likelihood values for each multistart run (if more than one multistart was used). \cr
   \code{time} \tab The user, system, and elapsed time to run the optimization. \cr
   \code{iterations} \tab The number of iterations until convergence. \cr
   \code{message} \tab A more informative message with the status of the optimization result. \cr
   \code{status} \tab An integer value with the status of the optimization (positive values are successes). Use \code{\link[=statusCodes]{statusCodes()}} for a detailed description. \cr
   \code{modelType} \tab The model type, \code{'mnl'} for multinomial logit or \code{'mxl'} for mixed logit. \cr
   \code{weightsUsed} \tab \code{TRUE} or \code{FALSE} for whether weights were used in the model. \cr
   \code{cluster} \tab The name of the column used to identify the cluster groups used in model estimation. \cr
   \code{numClusters} \tab The number of clusters. \cr
   \code{robust} \tab \code{TRUE} or \code{FALSE} for whether or not a robust covariance matrix was estimated. \cr
   \code{parSetup} \tab A summary of the distributional assumptions on each model parameter (\code{"f"}="fixed", \code{"n"}="normal distribution", \code{"ln"}="log-normal distribution"). \cr
   \code{standardDraws} \tab The draws used during maximum simulated likelihood (for MXL models). \cr
   \code{options} \tab A list of options for controlling the \code{nloptr()} optimization. Run \code{nloptr::nloptr.print.options()} for details. \cr
}
}
\description{
Use this function to estimate multinomial (MNL) and mixed logit (MXL)
models with "Preference" space or "Willingness-to-pay" (WTP) space utility
parameterizations. The function includes an option to run a multistart
optimization loop with random starting points in each iteration, which is
useful for non-convex problems like MXL models or models with WTP space
utility parameterizations. The main optimization loop uses the \code{nloptr()}
function to minimize the negative log-likelihood function.
}
\details{
The following options control the detailed behavior of the optimization
algorithm. They must be provided as a named list to the \code{options} argument,
e.g. \code{options = list(...)}. Below are a list of the default options.
Run \code{nloptr::nloptr.print.options()} for more options\tabular{lll}{
   Argument \tab Description \tab Default \cr
   \code{xtol_rel} \tab The relative \code{x} tolerance for the \code{nloptr} optimization loop. \tab \code{1.0e-6} \cr
   \code{xtol_abs} \tab The absolute \code{x} tolerance for the \code{nloptr} optimization loop. \tab \code{1.0e-6} \cr
   \code{ftol_rel} \tab The relative \code{f} tolerance for the \code{nloptr} optimization loop. \tab \code{1.0e-6} \cr
   \code{ftol_abs} \tab The absolute \code{f} tolerance for the \code{nloptr} optimization loop. \tab \code{1.0e-6} \cr
   \code{maxeval} \tab The maximum number of function evaluations for the \code{nloptr} optimization loop. \tab \code{1000} \cr
   \code{algorithm} \tab The optimization algorithm that \code{nloptr} uses. \tab \code{"NLOPT_LD_LBFGS"} \cr
   \code{print_level} \tab The print level of the \code{nloptr} optimization loop. \tab \code{0} \cr
}
}
\examples{
# For more detailed examples, visit
# https://jhelvy.github.io/logitr/articles/

library(logitr)

# Estimate a MNL model in the Preference space
mnl_pref <- logitr(
  data   = yogurt,
  choice = "choice",
  obsID  = "obsID",
  pars   = c("price", "feat", "brand")
)

# Estimate a MNL model in the WTP space, using a 10-run multistart
mnl_wtp <- logitr(
  data           = yogurt,
  choice         = "choice",
  obsID          = "obsID",
  pars           = c("feat", "brand"),
  price          = "price",
  modelSpace     = "wtp",
  numMultiStarts = 10
)
}
\keyword{logit}
\keyword{logitr}
\keyword{mixed}
\keyword{mnl}
\keyword{mxl}
\keyword{willingness-to-pay}
\keyword{wtp}
