% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{logitr}
\alias{logitr}
\title{The main function for running the logitr program}
\usage{
logitr(data, choiceName, obsIDName, betaNames, priceName = NULL,
  betaDist = NULL, priceDist = NULL, prefSpaceModel = NULL,
  standardDraws = NULL, options = list())
}
\description{
logitr estimates multinomial (MNL) and mixed logit (MXL) models in R with
options to estimate in the preference space or willingness-to-pay (WTP)
spaces. The program includes an option to run a multistart optimization
loop with random starting points to search for a global solution, which
should be done for WTP space models as well as mixed logit models since
they have nonlinear-in-parameters utility functions that may result in
multiple local maxima. The main optimization loop uses the nloptr function
to minimize the negative log-likelihood function.
}
\examples{
# Put example code here
}
\keyword{logitr,}
\keyword{mnl,}
\keyword{mxl}
