% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{miscmethods.logitr}
\alias{miscmethods.logitr}
\alias{logLik.logitr}
\alias{terms.logitr}
\alias{coef.logitr}
\alias{coef.summary.logitr}
\alias{summary.logitr}
\alias{print.logitr}
\alias{print.summary.logitr}
\alias{vcov.logitr}
\alias{print.logitr_wtp}
\title{Methods for logitr objects}
\usage{
\method{logLik}{logitr}(object, ...)

\method{terms}{logitr}(x, ...)

\method{coef}{logitr}(object, ...)

\method{coef}{summary.logitr}(object, ...)

\method{summary}{logitr}(object, ...)

\method{print}{logitr}(
  x,
  digits = max(3, getOption("digits") - 2),
  width = getOption("width"),
  ...
)

\method{print}{summary.logitr}(
  x,
  digits = max(3, getOption("digits") - 2),
  width = getOption("width"),
  ...
)

\method{vcov}{logitr}(object, ...)

\method{print}{logitr_wtp}(
  x,
  digits = max(3, getOption("digits") - 2),
  width = getOption("width"),
  ...
)
}
\arguments{
\item{object}{is an object of class \code{logitr}.}

\item{...}{further arguments.}

\item{x}{is an object of class \code{logitr}.}

\item{digits}{the number of digits for printing, defaults to \code{3}.}

\item{width}{the width of the printing,}
}
\description{
Miscellaneous methods for \code{logitr} class objects.
}
