% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.linearr}
\alias{predict.linearr}
\title{Predict Linear Regression}
\usage{
\method{predict}{linearr}(object, X, y = NULL)
}
\arguments{
\item{object}{'linearr' object or matrix of betas}

\item{X}{matrix or data frame of (new) observations}

\item{y}{optional, matrix or vector of response values}
}
\value{
predictions and loss metrics
}
\description{
Generates prediction for linear regression. Note that one can either input a 'linearr' object or a matrix of beta coefficients.
}
\examples{
library(dplyr)
X = dplyr::select(iris, -c(Species, Sepal.Length))
y = dplyr::select(iris, Sepal.Length)
fitted = linearr(X, y, lam = 0.1)
predict(fitted, X)
}
