% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.logisticr}
\alias{predict.logisticr}
\title{Predict Logistic Regression}
\usage{
\method{predict}{logisticr}(object, X, y = NULL)
}
\arguments{
\item{object}{'logisticr' object or matrix of betas}

\item{X}{matrix or data frame of (new) observations}

\item{y}{optional, matrix or vector of response values 0,1}
}
\value{
predictions and loss metrics
}
\description{
Generates prediction for logistic regression. Note that one can either input a 'logisticr' object or a matrix of beta coefficients.
}
\examples{
library(dplyr)
X = dplyr::select(iris, -Species)
y = dplyr::select(iris, Species)
y$Species = ifelse(y$Species == 'setosa', 1, 0)
logisticr(X, y)

fitted = logisticr(X, y, lam = 0.1, penalty = 'ridge', method = 'MM')
predict(fitted, X)
}
