% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{predictChoices}
\alias{predictChoices}
\title{Predict choices}
\usage{
predictChoices(
  model,
  alts,
  choiceName = NULL,
  obsIDName = NULL,
  priceName = NULL
)
}
\arguments{
\item{model}{The output of a model estimated model using the \code{logitr()}
function.}

\item{alts}{A data frame of a set of alternatives for which to predict
choices. Each row is an alternative and each column an attribute
corresponding to parameter names in the estimated model.}

\item{choiceName}{The name of the column that identifies the choice variable.
Include if you want to compare true choices from actual observations (e.g.
hold outs) to the predicted choices.}

\item{obsIDName}{The name of the column that identifies each set of
alternatives. Required if simulating results for more than one set of
alternatives. Defaults to \code{NULL} (for a single set of alternatives).}

\item{priceName}{The name of the parameter that identifies price. Only
required for WTP space models. Defaults to \code{NULL}.}
}
\value{
A data frame with the predicted choices for each alternative in
\code{alts}.
}
\description{
Returns the expected choices for a set of one or more alternatives based
on the results from an estimated model.
}
\examples{
\dontrun{
# Run a MNL model in the Preference Space:
library(logitr)

mnl_pref <- logitr(
  data = yogurt,
  choiceName = "choice",
  obsIDName = "obsID",
  parNames = c("price", "feat", "brand")
)

# Create a set of alternatives for which to simulate shares. Each row is an
# alternative and each column an attribute. In this example, I just use a
# couple of the choice observations from the yogurt dataset:
alts <- subset(yogurt, obsID \%in\% c(42, 13),
               select = c('price', 'feat', 'brand'))
alts$obsID <- c(rep(1, 4), rep(2, 4))
alts

# Predict choices using the estimated preference space MNL model:
predictChoices(mnl_pref, alts, obsIDName = "obsID")
}
}
\keyword{logitr}
\keyword{simluation}
