% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictChoices}
\alias{predictChoices}
\title{Predict choices}
\usage{
predictChoices(model, alts, altIDName, obsIDName = NULL)
}
\arguments{
\item{model}{The output of a model estimated model using the \code{logitr()}
function.
Include if you want to compare true choices from actual observations (e.g.
hold outs) to the predicted choices.}

\item{alts}{A data frame of a set of alternatives for which to predict
choices. Each row is an alternative and each column an attribute
corresponding to parameter names in the estimated model.}

\item{altIDName}{The name of the column that identifies each alternative
in each set of alternatives.}

\item{obsIDName}{The name of the column that identifies each set of
alternatives. Required if predicting results for more than one set of
alternatives. Defaults to \code{NULL} (for a single set of alternatives).}
}
\value{
A data frame with the predicted choices for each alternative in
\code{alts}.
}
\description{
Returns the expected choices for a set of one or more alternatives based
on the results from an estimated model.
}
\examples{
library(logitr)

# Estimate a preference space model
mnl_pref <- logitr(
  data       = yogurt,
  choiceName = "choice",
  obsIDName  = "obsID",
  parNames   = c("price", "feat", "brand")
)

# You can predict choices for any set of alternative, such as hold out
# samples or within-sample. For this example, choices will be predicted for
# the full yogurt data set, which was used to estimate the model.

# Predict choices using the estimated preference space MNL model
choices <- predictChoices(
  model     = mnl_pref,
  alts      = yogurt,
  altIDName = "alt",
  obsIDName = "obsID"
)

head(choices)

# Compute the accuracy
chosen <-  subset(choices, choice == 1)
chosen$correct <- chosen$choice == chosen$choice_predict
sum(chosen$correct) / nrow(chosen) # \% correctly predicted
}
\keyword{choice}
\keyword{logitr}
\keyword{predict}
\keyword{simulation}
