% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{runtimes}
\alias{runtimes}
\title{Data frame of run times for logitr benchmark}
\format{
\tabular{ll}{
   Variable \tab Description \cr
   \code{package} \tab Package name. \cr
   \code{time_sec} \tab The estimation time in seconds. \cr
   \code{numDraws} \tab The number of random draws used during estimation. \cr
}
}
\source{
\href{https://colab.research.google.com/drive/1vYlBdJd4xCV43UwJ33XXpO3Ys8xWkuxx?usp=sharing}{This} Google colab notebook
}
\usage{
data(runtimes)
}
\description{
This data frame contains the run times for a benchmark comparing the
relative computation time to estimate a preference space mixed logit model
using the following R packages: logitr, mixl, mlogit, gmnl, and
apollo. The run times are exported from the Google colab notebook here:
https://colab.research.google.com/drive/1vYlBdJd4xCV43UwJ33XXpO3Ys8xWkuxx?usp=sharing
}
\examples{
data(runtimes)

head(runtimes)
}
\keyword{datasets}
