% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSize.R
\name{sampleSizer}
\alias{sampleSizer}
\title{Examine the standard errors from models before you collect choice data}
\usage{
sampleSizer(data, obsIDName, parNames, nbreaks = 10, plot = TRUE,
  priceName = NULL, randPars = NULL, randPrice = NULL,
  modelSpace = "pref", options = list())
}
\description{
This function allows you to estimate a model on a design of experiment
that you have not yet used to collect data. This allows you to learn
about deficiencies in your design of experiment and also assess the sample
size needed to achieve parameter precision levels before you go out and
use the design to collect data. The function fills out the survey with 
random choices and estimates a model. It does this multiple times with an 
increasing number of observations, set by the "nbreaks" argument. While the 
coefficients in those models are meaningless, the standard errors on the 
coefficients are informative. The example below estimates 10 separate models 
and then plots the standard errors against the number of observations. In 
this example, assume that the yogurt data was not a completed survey but 
rather a blank design of experiment with no observed choices.
}
\examples{

test <- sampleSizer(
    data       = yogurt,
    obsIDName  = 'obsID',
    parNames   = c('price', 'feat', 'dannon', 'hiland', 'yoplait'), 
    nbreaks    = 10, 
    plot       = TRUE)
}
\keyword{logit,}
\keyword{logitr,}
\keyword{mnl,}
\keyword{mxl,}
\keyword{sample}
\keyword{size}
