% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{se.logitr}
\alias{se.logitr}
\title{Extract standard errors}
\usage{
\method{se}{logitr}(object, ...)
}
\arguments{
\item{object}{is an object of class \code{logitr} (a model estimated using
the 'logitr()` function).}

\item{...}{further arguments.}
}
\description{
Extract standard errors
}
