% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{simulateShares}
\alias{simulateShares}
\title{Returns the expected shares of a specific set of alternatives based
on an estimated model.}
\usage{
simulateShares(model, alts, priceName, alpha)
}
\description{
Returns the expected shares of a specific set of alternatives based
on an estimated model.
}
\examples{
# Run a MNL model in the Preference Space:
data(yogurt)

mnl.pref = logitr(
  data       = yogurt,
  choiceName = 'choice',
  obsIDName  = 'obsID',
  parNames   = c('price', 'feat', 'dannon', 'hiland', 'yoplait'))

# Create a set of alternatives for which to simulate shares:
alts = subset(yogurt, obsID==42,
       select=c('feat', 'price', 'dannon', 'hiland', 'yoplait'))
row.names(alts) = c('dannon', 'hiland', 'weight', 'yoplait')
alts

# Run the simulation using the estimated preference space MNL model:
simulateShares(mnl.pref, alts, alpha=0.025)
}
\keyword{logitr,}
\keyword{simluation}
