% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{statusCodes}
\alias{statusCodes}
\title{View a description the nloptr status codes}
\usage{
statusCodes()
}
\description{
Prints a description of the status codes from the nloptr optimization routine.
}
\details{
\tabular{ll}{
   Code \tab Description \cr
   1 \tab Generic success return value. \cr
   2 \tab Optimization stopped because stopval was reached. \cr
   3 \tab Optimization stopped because ftol_rel or ftol_abs was reached. \cr
   4 \tab Optimization stopped because xtol_rel or xtol_abs was reached. \cr
   5 \tab Optimization stopped because maxeval was reached. \cr
   6 \tab Optimization stopped because maxtime was reached. \cr
   -1 \tab Generic failure code. \cr
   -2 \tab Invalid arguments (e.g. lower bounds are bigger than upper bounds, \cr
   an unknown algorithm was specified, etc.). \tab  \cr
   -3 \tab Ran out of memory. \cr
   -4 \tab Halted because roundoff errors limited progress (in this case, the \cr
   optimization still typically returns a useful result.). \tab  \cr
   -5 \tab Halted because of a forced termination: the user called \cr
   \code{nlopt_force_stop(opt)} on the optimization's nlopt_opt object opt the \tab  \cr
   user's objective function or constraints. \tab  \cr
}
}
\examples{
statusCodes()
}
\keyword{codes}
\keyword{logitr,}
\keyword{nloptr,}
\keyword{status}
