% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.logitr}
\alias{summary.logitr}
\title{Prints a summary of an estimated model of the 'logitr' class.}
\usage{
\method{summary}{logitr}(model)
}
\description{
Prints a summary of an estimated model of the 'logitr' class.
}
\examples{
# View a summary of an estimate model:
summary(model)
}
\keyword{logitr,}
\keyword{summary}
