% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.logitr}
\alias{vcov.logitr}
\title{Calculate the variance-covariance matrix}
\usage{
\method{vcov}{logitr}(object, ...)
}
\arguments{
\item{object}{is an object of class \code{logitr}.}

\item{...}{further arguments.}
}
\description{
Returns the variance-covariance matrix of the main parameters of a fitted
model object.
}
