% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{wtpCompare}
\alias{wtpCompare}
\title{Returns a comparison of the WTP between a preference space and WTP space
model.}
\usage{
wtpCompare(model.pref, model.wtp, priceName)
}
\description{
Returns a comparison of the WTP between a preference space and WTP space
model.
}
\examples{
# Run a MNL model in the Preference Space:
data(yogurt)

mnl.pref = logitr(
  data       = yogurt,
  choiceName = 'choice',
  obsIDName  = 'obsID',
  parNames   = c('price', 'feat', 'dannon', 'hiland', 'yoplait'))

# Get the WTP implied from the preference space model
mnl.pref.wtp = wtp(mnl.pref, priceName='price')

# Run a MNL model in the WTP Space:
mnl.wtp = logitr(
  data       = yogurt,
  choiceName = 'choice',
  obsIDName  = 'obsID',
  parNames   = c('feat', 'dannon', 'hiland', 'yoplait'),
  priceName  = 'price',
  modelSpace = 'wtp',
  options = list(startVals = mnl.pref.wtp$Estimate))

# Compare the WTP between the two spaces:
wtpCompare(mnl.pref, mnl.wtp, priceName='price')
}
\keyword{logitr,}
\keyword{wtp}
